refactor(hash-aggr): consolidate aggregate hash table
