+++ datafusion/physical-plan/src/aggregates/aggregate_hash_table/mod.rs
@@ -17,14 +17,11 @@
 
 mod common;
 mod common_ordered;
-mod final_table;
 mod ordered_final_table;
 mod ordered_partial_table;
-mod partial_reduce_table;
-mod partial_table;
 
 pub(super) use common::{
-    AggregateHashTable, FinalMarker, PartialMarker, PartialReduceMarker,
-    PartialSkipMarker,
+    AggregateHashTable, AggregateTableMode, FinalMarker, PartialMarker,
+    PartialSkipHashTable,
 };
 pub(super) use common_ordered::OrderedAggregateTable;
